moved around the structure slightly to make certain files more acccessible; also edited the routes for server.js
